LayoutTests/media/modern-media-controls/placard-support/placard-support-airplay.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<body>
<style type="text/css" media="screen">
    
    video, #host {
        position: absolute;
        top: 0;
        left: 0;
    }

    video {
        width: 800px;
        height: 240px;
    }
    
</style>
<video src="../../content/test.mp4" autoplay></video>
<div id="host"></div>
<script type="text/javascript">

window.jsTestIsAsync = true;

description("Testing the <code>PlacardSupport</code> behavior when playing media through AirPlay.");

iconService.directoryPath = "../../../../Source/WebCore/Modules/modern-media-controls/images";

const container = document.querySelector("div#host");
const media = document.querySelector("video");
const mediaController = createControls(container, media, null);

if (window.internals) {
    window.internals.settings.setAllowsAirPlayForMediaPlayback(true);
    window.internals.setMediaElementRestrictions(media, "norestrictions");
}

media.addEventListener("webkitcurrentplaybacktargetiswirelesschanged", () => {
    if (!media.paused && media.webkitCurrentPlaybackTargetIsWireless) {
        shouldBeTrue("mediaController.controls.showsPlacard");
        shouldBeTrue("mediaController.controls.children.includes(mediaController.controls.airplayPlacard)");

        container.remove();
        media.remove();
        finishJSTest();
    }
});

media.addEventListener("play", () => {
    media.webkitShowPlaybackTargetPicker();
    if (window.internals)
        window.internals.setMockMediaPlaybackTargetPickerState("Sleepy TV", "DeviceAvailable");
});

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
