LayoutTests/media/modern-media-controls/placard-support/placard-support-airplay-expected.txt

Testing the PlacardSupport behavior when playing media through AirPlay.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


got event false
got event true
PASS mediaController.controls.showsPlacard is true
PASS mediaController.controls.children.includes(mediaController.controls.airplayPlacard) is true
PASS successfullyParsed is true

TEST COMPLETE
